+++ docs/ServiceNow-connector.md
@@ -1,0 +1,28 @@
+# ServiceNow Connection
+
+Description
+-----------
+Use this connection to access data in ServiceNow.
+
+Properties
+----------
+**Name:** Name of the connection. Connection names must be unique in a namespace.
+
+**Description:** Description of the connection.
+
+**Client ID**: The Client ID for ServiceNow Instance.
+
+**Client Secret**: The Client Secret for ServiceNow Instance.
+
+**REST API Endpoint**: The REST API Endpoint for ServiceNow Instance. 
+
+**User Name**: The user name for ServiceNow Instance.
+
+**Password**: The password for ServiceNow Instance.
+
+
+Path of the connection
+----------------------
+To browse, get a sample from, or get the specification for this connection (Not supported in ServiceNow Batch Multi 
+Source plugins).  
+/{table} This path indicates a ServiceNow table. A table is the only one that can be sampled.
