+++ docs/ServiceNowMultiSource-batchsource.md
@@ -8,19 +8,24 @@ Reads from one or multiple tables within ServiceNow
 Properties
 ----------
 
-**Reference Name**: Name used to uniquely identify this source for lineage, annotating metadata, etc.
+**Use Connection:** Whether to use a connection. If a connection is used, you do not need to provide the credentials.
 
-**Table Names**: The name of the ServiceNow table(s) from which data to be fetched.
+**Connection:** Name of the connection to use. You also can use the macro function ${conn(connection-name)}.  
 
 **Client ID**: The Client ID for ServiceNow Instance.
 
 **Client Secret**: The Client Secret for ServiceNow Instance.
 
-**REST API Endpoint**: The REST API Endpoint for ServiceNow Instance. For example, `https://instance.service-now.com`
+**REST API Endpoint**: The REST API Endpoint for ServiceNow Instance. 
 
 **User Name**: The user name for ServiceNow Instance.
 
 **Password**: The password for ServiceNow Instance.
+
+**Reference Name**: Name used to uniquely identify this source for lineage, annotating metadata, etc.
+
+**Table Names**: The name of the ServiceNow table(s) from which data to be fetched.
+
 
 **Start Date**: The Start date to be used to filter the data. The format must be `yyyy-MM-dd`.
